+++ app/Http/Controllers/MenusController.php
@@ -1,13 +1,4 @@
 <?php
-/*
-|--------------------------------------------------------------------------
-|MenusController
-|--------------------------------------------------------------------------
-|
-| This controller handles the  rooms check availability part for
-| each room types.
-|
-*/
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
@@ -15,21 +6,51 @@ use Illuminate\Http\Request;
 use App\Http\Requests;
 use App\Http\Controllers\Controller;
 use DB;
+
+/**
+ * Class MenusController
+ * @package App\Http\Controllers
+ */
 class MenusController extends Controller
 {
 
-    //Returns the Menus view
+    /*
+    |--------------------------------------------------------------------------
+    | Menus Controller
+    |--------------------------------------------------------------------------
+    |
+    |This controller provides views and feeds data to the admin's control
+    |panel's Menus Management section.
+    |
+    */
+
+    /**
+     * Returns the Menus view.
+     *
+     * @param Request $request
+     * @return \Illuminate\Contracts\View\Factory|\Illuminate\View\View
+     */
     public function getIndex (Request $request){
         return view('nipuna.menus');   
     }
 
-    //Returns all the Menus from the DB for the DataTable
+    /**
+     * Returns all the Menus from the DB for the DataTable.
+     *
+     * @param Request $request
+     * @return \Illuminate\Http\JsonResponse
+     */
     public function getMenus(Request $request){
         $result = DB::table('menus')->get();
         return response()->json(['count' => count($result), 'data' => $result]);
     }
 
-    //Returns all the items for a given row number of a Menu
+    /**
+     * Returns all the items for a given row number of a Menu.
+     *
+     * @param Request $request
+     * @return mixed
+     */
     public function getMenuitemdataload(Request $request){
         $rowno = $request->input('rowno');
 
@@ -37,7 +58,12 @@ class MenusController extends Controller
         return $result;
     }
 
-    //Returns the details of a Menu when the row number is given
+    /**
+     * Returns the details of a Menu when the row number is given.
+     *
+     * @param Request $request
+     * @return mixed
+     */
     public function getRetrievedetails(Request $request){
         $menu_id=$request->input('row');
 
@@ -45,7 +71,12 @@ class MenusController extends Controller
         return $result;
     }
 
-    //Returns all the items in a Menu when the row number is given
+    /**
+     * Returns all the items in a Menu when the row number is given.
+     *
+     * @param Request $request
+     * @return mixed
+     */
     public function getRetrieveitemdetails(Request $request){
         $row=$request->input('row');
 
@@ -53,7 +84,12 @@ class MenusController extends Controller
         return $result;
     }
 
-    //Returns all the items in a Menu in JSON format when the row number is given
+    /**
+     * Returns all the items in a Menu in JSON format when the row number is given.
+     *
+     * @param Request $request
+     * @return \Illuminate\Http\JsonResponse
+     */
     public function getDetailedmenus(Request $request){
         $row = $request->input('row');
 
@@ -61,14 +97,24 @@ class MenusController extends Controller
         return response()->json(['count' => count($result), 'data' => $result]);
     }
 
-    //Returns the number of items in a menu
+    /**
+     * Returns the number of items in a menu.
+     *
+     * @param Request $request
+     * @return mixed
+     */
     public function getMenuitemcount(Request $request){
         $menu_id = $request->input('menu_id');
         $result = DB::table('DETAILED_MENUS')->where('menu_id',$menu_id)->count();
         return $result;
     }
 
-    //Insert a new Menu and return the row number of the new entry
+    /**
+     * Insert a new Menu and return the row number of the new entry.
+     *
+     * @param Request $request
+     * @return mixed
+     */
     public function getAddmenu(Request $request){
         $category = $request->input('category');
         $description = $request->input('description');
@@ -78,7 +124,12 @@ class MenusController extends Controller
         return $rowno;
     }
 
-    //Insert a new item for the Menu
+    /**
+     * Insert a new item for the Menu.
+     *
+     * @param Request $request
+     * @return int
+     */
     public function getAddmenuitem(Request $request){
         $menu_number = $request->input('menu_number');
         $item_name = $request->input('item_name');
@@ -88,7 +139,11 @@ class MenusController extends Controller
         return 1;
     }
 
-    //Update an item of the Menu
+    /**
+     * Update an item of the Menu.
+     *
+     * @param Request $request
+     */
     public  function getUpdateitem(Request $request){
         $rownumber = $request->input('id');
         $item_name = $request->input('item_name');
@@ -97,7 +152,11 @@ class MenusController extends Controller
         DB::table('DETAILED_MENUS')->where('id',$rownumber)->update(array('item_name'=> $item_name,'price'=> $price));
     }
 
-    //Update an existing Menu
+    /**
+     * Update an existing Menu.
+     *
+     * @param Request $request
+     */
     public function getUpdatemenu(Request $request){
         $category = $request->input('category');
         $description = $request->input('description');
@@ -106,8 +165,12 @@ class MenusController extends Controller
 
         DB::table('menus')->where('menu_id',$row)->update(array('category'=> $category,'description'=> $description,'rate'=> $rate));
     }
-    
-    //Delete the Menu when the row number is given
+
+    /**
+     * Delete the Menu when the row number is given.
+     *
+     * @param Request $request
+     */
     public function getDeleterow(Request $request){
         $menu_id=$request->input('row');
         DB::table('menus')->where('menu_id','=',$menu_id)->delete();
@@ -115,7 +178,11 @@ class MenusController extends Controller
 
     }
 
-    //Delete an item when the row number is given
+    /**
+     * Delete an item when the row number is given.
+     *
+     * @param Request $request
+     */
     public function getDeleteitem(Request $request){
         $item_id=$request->input('row');
         DB::table('DETAILED_MENUS')->where('id','=',$item_id)->delete();
